requirements.txt
environs==9.5.0
python-telegram-bot==13.13
notifiers==1.3.3
redis==4.3.4
requests==2.28.1
python-slugify==6.1.2
geopy==2.2.0
